Updated jsonl, made some other fixes
