+++ src/backend/bedrock_evaluations_runner/.env-example
@@ -1,6 +1,7 @@
-AWS_REGION="eu-west-2"
-ROLE_ARN="arn:aws:iam::767397886959:role/nhs-REPLACE-notifyai-bedrock-automatic-evaluation-role"
-INPUT_PROMPT_S3_URI="s3://nhs-REPLACE-notifyai-input-prompts/prompts.jsonl"
-RESULTS_S3_URI="s3://nhs-REPLACE-notifyai-results/results/"
-EVALUATOR_MODEL_IDENTIFIER="amazon.nova-pro-v1:0"
-INFERENCE_MODEL_IDENTIFIER="amazon.nova-pro-v1:0"
+env_region="eu-west-2"
+env_role_arn="arn:aws:iam::767397886959:role/nhs-REPLACE-notifyai-bedrock-automatic-evaluation-role"
+env_input_prompt_s3_uri="s3://nhs-REPLACE-notifyai-input-prompts/prompts.jsonl"
+env_results_s3_uri="s3://nhs-REPLACE-notifyai-results/results/"
+env_evaluator_model_identifier="amazon.nova-pro-v1:0"
+env_generator_model_identifier="amazon.nova-pro-v1:0"
+env_resource_prefix="nhs-REPLACE-notifyai"
