+++ tools/prepare_training_data.py/prepare_training_data.py
