+++ tools/prompt-input-generator/env.example
@@ -1,0 +1,5 @@
+LETTERS_PATH = PATH-TO-LETTERS-YOU-WISH-TO-EVALUATE-WITH
+RESPONSE_PATH = infrastructure/terraform/components/notifyai/resources/prompt-data/letter-responses.json
+DATA_PATH = infrastructure/terraform/components/notifyai/resources/prompt-data/prompts.jsonl
+MODEL = amazon.nova-pro-v1:0
+PROMPT = src/backend/bedrock-prompt-messager/system_prompt.txt
